frontend: Reorganise components directory & refactor imports
