+++ frontend/src/common/components/authentication/Logout.tsx
