+++ frontend/src/common/components/authentication/index.ts
@@ -1,0 +1,4 @@
+import Logout from "./Logout";
+import Unauthorized from "./Unauthorized";
+
+export {Logout, Unauthorized};
