+++ frontend/src/features/housingCompany/HousingCompanyListPage.tsx
@@ -4,15 +4,13 @@ import {Button, IconPlus, IconSearch, LoadingSpinner, StatusLabel} from "hds-rea
 import {Link} from "react-router-dom";
 
 import {useGetDevelopersQuery, useGetHousingCompaniesQuery, useGetPropertyManagersQuery} from "../../app/services";
+import {Heading, ListPageNumbers, QueryStateHandler} from "../../common/components";
 import {
     FilterIntegerField,
+    FilterRelatedModelComboboxField,
     FilterSelectField,
     FilterTextInputField,
-    Heading,
-    ListPageNumbers,
-    QueryStateHandler,
-} from "../../common/components";
-import FilterRelatedModelComboboxField from "../../common/components/FilterRelatedModelComboboxField";
+} from "../../common/components/filters";
 import {getHousingCompanyHitasTypeName} from "../../common/localisation";
 import {IHousingCompany, IHousingCompanyListResponse} from "../../common/schemas";
 import {formatDate} from "../../common/utils";
